+++ ROCKET_DESIGN_FILES/Nose_Cone/Fusion360_Script/.env
@@ -1,0 +1,1 @@
+PYTHONPATH=/Users/varunahlawat/Library/Application Support/Autodesk/webdeploy/production/e379c5f768f16e4cbcd81b81fe59a8562f304f19/Api/Python/packages
