+++ ROCKET_DESIGN_FILES/Nose_Cone/Fusion360_Script/.vscode/launch.json
@@ -1,0 +1,20 @@
+{
+	"version":	"0.2.0",
+	"configurations":	[{
+			"name":	"Python: Attach",
+			"type":	"python",
+			"request":	"attach",
+			"pathMappings":	[{
+					"localRoot":	"${workspaceRoot}",
+					"remoteRoot":	"${workspaceRoot}"
+				}],
+			"osx":	{
+				"filePath":	"${file}"
+			},
+			"windows":	{
+				"filePath":	"${file}"
+			},
+			"port":	9000,
+			"host":	"localhost"
+		}]
+}
