+++ ROCKET_DESIGN_FILES/Nose_Cone/Fusion360_Script/.vscode/settings.json
@@ -1,0 +1,5 @@
+{
+	"python.autoComplete.extraPaths":	["/Users/varunahlawat/Library/Application Support/Autodesk/Autodesk Fusion 360/API/Python/defs"],
+	"python.analysis.extraPaths":	["/Users/varunahlawat/Library/Application Support/Autodesk/Autodesk Fusion 360/API/Python/defs"],
+	"python.defaultInterpreterPath":	"/Users/varunahlawat/Library/Application Support/Autodesk/webdeploy/production/e379c5f768f16e4cbcd81b81fe59a8562f304f19/Autodesk Fusion 360.app/Contents/Frameworks/Python.framework/Versions/Current/bin/python"
+}
