+++ ROCKET_DESIGN_FILES/Nose_Cone/Fusion360_Script/Haack Series NoseCone.manifest
@@ -1,0 +1,10 @@
+{
+	"autodeskProduct":	"Fusion360",
+	"type":	"script",
+	"author":	"Varun Ahlawat",
+	"description":	{
+		"":	"Haack gave a formula for nose cones for least amount of aerodynamic drag for a fixed volume and a fixed diameter. His formula goes like this:\n\n# Function to create a cone based on Haack series\n    # Create a new sketch on the XY plane\n    sketches = root_comp.sketches\n    xy_plane = root_comp.xYConstructionPlane\n    sketch = sketches.add(xy_plane)\n\n    # Calculate the radius at each point along the Haack series\n    def haack_series(radius, c):\n        return radius * math.sqrt((c * 2 * math.pi) / (2 * math.pi * radius))\n\n    # Parameters for the Haack series\n    base_radius = 1.0\n    segments = 100\n\n    # Create the Haack series profile\n    points = []\n    for i in range(segments + 1):\n        angle = (2 * math.pi * i) / segments\n        radius = base_radius * haack_series(math.cos(angle), c_value)\n        points.append(adsk.core.Point3D.create(radius * math.cos(angle), radius * math.sin(angle), 0)"
+	},
+	"supportedOS":	"windows|mac",
+	"editEnabled":	true
+}
